Update docker-compose entry point

+++ docker-compose.web.sh
@@ -1,3 +1,14 @@
 #!/usr/bin/env sh
-rm -rf .cache/*
+
+# Yarn install as needed
+if [ ! -x ./node_modules/.bin/gatsby ]; then
+  yarn
+fi
+
+# Clear cache as needed
+if [ -d .cache ]; then
+  rm -rf .cache/*
+fi
+
+# Yarn run
 yarn run gatsby develop -- --host 0.0.0.0 --port 25800
